Testimonial delete endpoint
