+++ src/main/java/com/alkemy/ong/controller/TestimonialController.java
@@ -4,14 +4,11 @@ import java.net.URI;
 import com.alkemy.ong.dto.TestimonialDTO;
 import com.alkemy.ong.service.TestimonialService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 @RestController
@@ -51,5 +48,17 @@ public class TestimonialController extends BaseController {
     return ResponseEntity.ok(testimonialService.update(id, testimonial));
   }
 
+  // --------------------------------------------------------------------------------------------
+  // Delete
+  // --------------------------------------------------------------------------------------------
 
+  @PreAuthorize("ROL_ADMIN")
+  @DeleteMapping("/{id}")
+  public ResponseEntity<Void> delete(@PathVariable String id) {
+    if (testimonialService.findById(id) == null) {
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+    }
+    testimonialService.delete(id);
+    return ResponseEntity.status(HttpStatus.OK).build();
+  }
 }
