+++ src/main/java/com/alkemy/ong/service/TestimonialService.java
@@ -8,50 +8,14 @@ import com.alkemy.ong.repository.TestimonialRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-@Service
-public class TestimonialService {
+public interface TestimonialService {
 
-  // --------------------------------------------------------------------------------------------
-  // Autowireds
-  // --------------------------------------------------------------------------------------------
+  TestimonialDTO create(TestimonialDTO dto);
 
-  @Autowired
-  private TestimonialRepository testimonialRepository;
+  TestimonialDTO findById(String id);
 
-  @Autowired
-  private TestimonialMapper testimonialMapper;
+  TestimonialDTO update(String id, TestimonialDTO dto);
 
-  // --------------------------------------------------------------------------------------------
-  // Create
-  // --------------------------------------------------------------------------------------------
-
-  public TestimonialDTO create(TestimonialDTO dto) {
-    Testimonial testimonial = testimonialRepository.save(testimonialMapper.toEntity(dto));
-    return testimonialMapper.toDTO(testimonial);
-  }
-
-  // --------------------------------------------------------------------------------------------
-  // Read
-  // --------------------------------------------------------------------------------------------
-
-  // --------------------------------------------------------------------------------------------
-  // Update
-  // --------------------------------------------------------------------------------------------
-
-  public TestimonialDTO update(String id, TestimonialDTO dto) {
-
-    Testimonial testimonial =
-        testimonialRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-
-    testimonial.setName(dto.getName());
-    testimonial.setContent(dto.getImage());
-    testimonial.setImage(dto.getImage());
-
-    return testimonialMapper.toDTO(testimonialRepository.save(testimonial));
-  }
-
-  // --------------------------------------------------------------------------------------------
-  // Delete
-  // --------------------------------------------------------------------------------------------
+  void delete(String id);
 
 }
